@@ -1,10 +1,10 @@
 name: dispose
 description: A handler to correctly dispose listeners, timers, controllers etc.
 homepage: https://github.com/jodinathan/dispose
-version: 0.0.5
+version: 0.0.6
 
 environment:
-  sdk: '> 2.4.0 < 3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  pedantic: ^1.9.2
+  lints: ^1.0.1
